remove metric

+++ countrymaam.go
@@ -7,6 +7,6 @@ import (
 
 type Index[T number.Number, U any] interface {
 	Add(feature []T, item U)
-	Search(feature []T, n uint, r float32) ([]index.Candidate[U], error)
+	Search(feature []T, n uint, r float64) ([]index.Candidate[U], error)
 	Build() error
 }
