+++ index/common.go
@@ -1,6 +1,6 @@
 package index
 
 type Candidate[U any] struct {
-	Distance float32
+	Distance float64
 	Item     U
 }
